Update index.html
updated with interactivity for the inline data viz. Note: weird bug with the time slider where when set to zero the slider has the right value but does not display it properly on the slider. Im not sure why and I have yet to be able to fix it. I will update this file if I find a solution.